+++ pytest_inmanta/core.py
@@ -37,6 +37,11 @@ except DistributionNotFound:
 
 # Setting a project-wide pip index is only supported for iso7+
 SUPPORTS_PROJECT_PIP_INDEX: bool = (
+    CORE_VERSION is not None and CORE_VERSION >= version.Version("11.0.0.dev")
+)
+
+
+SUPPORTS_LEGACY_PROJECT_PIP_INDEX: bool = (
     CORE_VERSION is not None and CORE_VERSION >= version.Version("9.0.0.dev")
 )