+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v 2.9.0 (2023-11-29)
+Changes in this release:
+- Add support for more values (yes, 1) and (no, 0) for boolean options
+- Add `deploy_resouce_v2` endpoint
+- Add support for unmanaged resources
+- Adds a new resource entity that raises an IgnoreResourceException when buildings its id
+- Add support for ISO7. 
+- Add `--pip-pre`, `--pip-use-system-config` and `--pip-index-url`
+
 # v 2.8.0 (2023-08-03)
 Changes in this release:
 - Log a warning when the id_attribute of a resource is called id.